improve pointer passing to thread

+++ hello_world/src/hello_world_pthreads.c
@@ -1,12 +1,13 @@
 #include <stdio.h>
 #include <stdlib.h>
+#include <stdint.h>
 #include <assert.h>
 #include <pthread.h>
 
 void *print_msg(void *thread_id)
 {
   int th_id;
-  th_id = (int)thread_id;
+  th_id = (intptr_t)thread_id;
   printf("Hello, World from thread %d\n", th_id);
   pthread_exit(NULL);
 }
@@ -20,7 +21,7 @@ int main (int argc, char *argv[])
   // create threads individually
   for(i=0; i<nthreads; i++)
     {
-      rc = pthread_create(&threads[i], NULL, print_msg, (void *) i);
+      rc = pthread_create(&threads[i], NULL, print_msg, (void *) (intptr_t) i);
       assert(rc == 0);
     }
